+++ src/server/auth-client.ts
@@ -125,7 +125,6 @@ import {
   DEFAULT_SCOPES,
   DEFAULT_STT_SCOPES
 } from "../utils/constants.js";
-import { withDPoPNonceRetry } from "../utils/dpopRetry.js";
 import { createSizeLimitedFetch } from "../utils/fetchUtils.js";
 import { createAuthCompletePostMessageResponse } from "../utils/html-helpers.js";
 import { buildEnrollOptions } from "../utils/mfa-server-utils.js";
@@ -191,6 +190,7 @@ import {
   type ReadonlyRequestCookies
 } from "./cookies.js";
 import { DiscoveryCache } from "./discovery-cache.js";
+import { withDPoPNonceRetry } from "./dpop/retry.js";
 import {
   AccessTokenFactory,
   Fetcher,
@@ -639,7 +639,7 @@ export class AuthClient {
     }
 
     // Dynamic import only when needed - prevents crypto from being bundled
-    const dpopModule = await import("../utils/dpopUtils.js");
+    const dpopModule = await import("./dpop/utils.js");
     const dpopConfig = await dpopModule.validateDpopConfiguration({
       useDPoP: this.useDPoP,
       dpopKeyPair: this.dpopKeyPair, // Pass existing keypair for validation
