+++ src/server/dpop/dpop-authcode-nonce-retry.test.ts
@@ -8,9 +8,9 @@ import { beforeAll, describe, expect, it } from "vitest";
 import { getDefaultRoutes } from "../../test-fixtures/defaults.js";
 import { generateSecret } from "../../test-fixtures/utils.js";
 import { RESPONSE_TYPES, TransactionState } from "../../types/index.js";
-import { generateDpopKeyPair } from "../../utils/dpopRetry.js";
 import { AuthClient } from "../auth-client.js";
 import { decrypt, encrypt } from "../cookies.js";
+import { generateDpopKeyPair } from "../dpop/retry.js";
 import { StatelessSessionStore } from "../session/stateless-session-store.js";
 import { TransactionStore } from "../transaction-store.js";
