+++ src/server/dpop/retry.ts
@@ -1,6 +1,6 @@
 import { generateKeyPair, isDPoPNonceError } from "oauth4webapi";
 
-import { DpopKeyPair, RetryConfig } from "../types/dpop.js";
+import { DpopKeyPair, RetryConfig } from "../../types/dpop.js";
 
 /**
  * Detects if the current environment is Edge Runtime.
