+++ src/server/index.ts
@@ -7,7 +7,7 @@ export { AbstractSessionStore } from "./session/abstract-session-store.js";
 export { filterDefaultIdTokenClaims, DEFAULT_ID_TOKEN_CLAIMS } from "./user.js";
 
 // DPoP utilities
-export { generateDpopKeyPair } from "../utils/dpopRetry.js";
+export { generateDpopKeyPair } from "./dpop/retry.js";
 
 // MFA error classes for handling MFA step-up authentication
 export {
