+++ src/server/my-account/proxy-handler.test.ts
@@ -16,8 +16,8 @@ import {
   extractDPoPInfo
 } from "../../test-fixtures/proxy-handler-test-helpers.js";
 import { generateSecret } from "../../test-fixtures/utils.js";
-import { generateDpopKeyPair } from "../../utils/dpopRetry.js";
 import { AuthClient } from "../auth-client.js";
+import { generateDpopKeyPair } from "../dpop/retry.js";
 import { StatelessSessionStore } from "../session/stateless-session-store.js";
 import { TransactionStore } from "../transaction-store.js";
