+++ src/server/passwordless/passwordless-db-server.flow.test.ts
@@ -10,8 +10,8 @@ import {
   setupMswLifecycle
 } from "../../test-fixtures/defaults.js";
 import { generateSecret } from "../../test-fixtures/utils.js";
-import { generateDpopKeyPair } from "../../utils/dpopRetry.js";
 import { AuthClient } from "../auth-client.js";
+import { generateDpopKeyPair } from "../dpop/retry.js";
 import { StatelessSessionStore } from "../session/stateless-session-store.js";
 import { TransactionStore } from "../transaction-store.js";
