+++ src/server/passwordless/passwordless-server.flow.test.ts
@@ -12,10 +12,10 @@ import {
 } from "../../test-fixtures/defaults.js";
 import { generateSecret, stripTransactionValuePrefix } from "../../test-fixtures/utils.js";
 import type { SessionData } from "../../types/index.js";
-import { generateDpopKeyPair } from "../../utils/dpopRetry.js";
 import { AuthClientProvider } from "../auth-client-provider.js";
 import { AuthClient } from "../auth-client.js";
 import { decrypt, encrypt } from "../cookies.js";
+import { generateDpopKeyPair } from "../dpop/retry.js";
 import { ServerPasswordlessClient } from "../passwordless/server-passwordless-client.js";
 import { StatelessSessionStore } from "../session/stateless-session-store.js";
 import { TransactionStore } from "../transaction-store.js";
