+++ src/server/tests/custom-token-exchange.test.ts
@@ -21,8 +21,8 @@ import { getDefaultRoutes } from "../../test-fixtures/defaults.js";
 import { generateSecret } from "../../test-fixtures/utils.js";
 import { SessionData } from "../../types/index.js";
 import { TOKEN_TYPES } from "../../types/token-vault.js";
-import { generateDpopKeyPair } from "../../utils/dpopRetry.js";
 import { AuthClient } from "../auth-client.js";
+import { generateDpopKeyPair } from "../dpop/retry.js";
 import { StatelessSessionStore } from "../session/stateless-session-store.js";
 import { TransactionStore } from "../transaction-store.js";
